# 8261553: Efficient mask generation using BMI2 BZHI instruction