--- a/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
+++ b/src/hotspot/cpu/x86/c2_MacroAssembler_x86.cpp
@@ -1894,17 +1894,9 @@ void C2_MacroAssembler::reduce8L(int opcode, Register dst, Register src1, XMMReg
 }
 
 void C2_MacroAssembler::genmask(Register dst, Register len, Register temp) {
-  if (ArrayCopyPartialInlineSize <= 32) {
-    mov64(dst, 1);
-    shlxq(dst, dst, len);
-    decq(dst);
-  } else {
-    mov64(dst, -1);
-    movq(temp, len);
-    negptr(temp);
-    addptr(temp, 64);
-    shrxq(dst, dst, temp);
-  }
+  assert(ArrayCopyPartialInlineSize <= 64,"");
+  mov64(dst, -1L);
+  bzhiq(dst, dst, len);
 }
 #endif // _LP64