BMI2 BHZI instruction can be used to optimize the instruction sequence
used for mask generation at various place in array copy stubs and partial in-lining for small copy operations.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8261553](https://bugs.openjdk.java.net/browse/JDK-8261553): Efficient mask generation using BMI2 BZHI instruction


### Reviewers
 * [Claes Redestad](https://openjdk.java.net/census#redestad) (@cl4es - **Reviewer**)
 * [Nils Eliasson](https://openjdk.java.net/census#neliasso) (@neliasso - **Reviewer**) ⚠️ Review applies to 84c9c2dab1bd5b9da846c4fe703a317269eb4839


### Download
`$ git fetch https://git.openjdk.java.net/jdk pull/2522/head:pull/2522`
`$ git checkout pull/2522`