@jatin-bhateja 2021-02-11
Hi Claes,

Here is the JMH performance data over CLX for arraycopy benchmarks:
http://cr.openjdk.java.net/~jbhateja/8261553/JMH_PERF_CLX_BASELINE.txt
http://cr.openjdk.java.net/~jbhateja/8261553/JMH_PERF_CLX_WITH_OPTS.txt

Regards,
Jatin