@jatin-bhateja 2021-02-11
> * Rather than removing the old code, I believe the code calling bzhiq needs to be in a branch checking `VM_Version::supports_bmi2`. Otherwise you'll hit asserts on older hardware without this extension

Hi Claes, added missing safely check for BMI2,  its in general  rare that a target supporting AVX-512 does not support BMI2

> * Some demonstration of the performance benefit would be nice - either a new microbenchmark or a statistically significant result running some existing ones, e.g. `make test TEST=micro:ArrayCopy`