@cl4es 2021-02-11
> Hi Claes,
> 
> Here is the JMH performance data over CLX for arraycopy benchmarks:
> http://cr.openjdk.java.net/~jbhateja/8261553/JMH_PERF_CLX_BASELINE.txt
> http://cr.openjdk.java.net/~jbhateja/8261553/JMH_PERF_CLX_WITH_OPTS.txt
> 
> Regards,
> Jatin

Thanks! Eyeballing the results it looks like a mixed bag. There even seems to be a few regressions such as this: 

```
o.o.b.java.lang.ArrayCopyUnalignedSrc.testLong                   1200     N/A   avgt    2     61.663           ns/op
--> 
o.o.b.java.lang.ArrayCopyUnalignedSrc.testLong                   1200     N/A   avgt    2     74.160           ns/op
```