@jatin-bhateja 2021-02-11
> > Hi Claes,
> > Here is the JMH performance data over CLX for arraycopy benchmarks:
> > http://cr.openjdk.java.net/~jbhateja/8261553/JMH_PERF_CLX_BASELINE.txt
> > http://cr.openjdk.java.net/~jbhateja/8261553/JMH_PERF_CLX_WITH_OPTS.txt
> > Regards,
> > Jatin
> 
> Thanks! Eyeballing the results it looks like a mixed bag. There even seems to be a few regressions such as this:
> 
> ```
> o.o.b.java.lang.ArrayCopyUnalignedSrc.testLong                   1200     N/A   avgt    2     61.663           ns/op
> --> 
> o.o.b.java.lang.ArrayCopyUnalignedSrc.testLong                   1200     N/A   avgt    2     74.160           ns/op
> ```

Hi Claes, This could be a run to run variation, in general we are now having fewer number of instructions (one shift operation saved per mask computation) compared to previous masked generation sequence  and thus it will always offer better execution latencies.